requirements.txt
anyio==4.8.0
asyncio==3.4.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
tzdata==2025.1
urllib3==2.3.0
yt-dlp==2025.1.26
